64-bit transition part 9: ksh93 command line editors (#960)
This is the ninth of the thickfold patch series, which enables
ksh93 to operate within a 64-bit address space.

The parts of ksh93 affected by this commit are:
- The interactive emacs and vi editor modes. Neither possess an
  exigent need to operate on exceedingly large buffers (I'd be
  surprised if they ever needed to operate in gigabytes to begin
  with), so I opted to keep many variables ints since those are
  more favorable for struct size. (Decreasing the sizes of the
  global editor structs would require reordering the variables in
  those structs, which is best left for a separate patch.)
  - Moved a FALLTHROUGH comment to fix a fallthrough warning.
- Various fixes for jobs.c.
  - Get rid of if/else PID botch; a single casted strtoll is fine.

Change in the number of warnings on Linux when compiling with clang
using -Wsign-compare -Wshorten-64-to-32 -Wsign-conversion
-Wimplicit-int-conversion: 1,032 => 911 => 37 (progression from
part 8 => part 9 => part 13)

Progresses https://github.com/ksh93/ksh/issues/592

+++ src/cmd/ksh93/include/edit.h
@@ -126,7 +126,7 @@ typedef struct edit
 	void	*e_vi;		/* vi specific data */
 	void	*e_emacs;	/* emacs specific data */
 	char	*e_stkptr;	/* saved stack pointer */
-	int	e_stkoff;	/* saved stack offset */
+	ptrdiff_t e_stkoff;	/* saved stack offset */
 	char	**e_clist;	/* completion list after <ESC>= */
 	int	e_nlist;	/* number of elements on completion list */
 #if SHOPT_ESH || SHOPT_VSH
@@ -189,8 +189,8 @@ extern void	*ed_open(void);
 	extern int ed_internal(const char*, genchar*);
 	extern int ed_external(const genchar*, char*);
 	extern void ed_gencpy(genchar*,const genchar*);
-	extern void ed_genncpy(genchar*,const genchar*,int);
-	extern int ed_genlen(const genchar*);
+	extern void ed_genncpy(genchar*,const genchar*,size_t);
+	extern size_t ed_genlen(const genchar*);
 #endif /* SHOPT_MULTIBYTE */
 #if SHOPT_EDPREDICT
     extern int	ed_histgen(Edit_t*, const char*);
